Basic Data

Basic types to contain the data for model training. This module defines the basic DataBunch struct that is used inside Learner to train a model. The fields are generic and can take any kind of fastai dataset (see IterableDataset and MapDataset) or Flux.Data.DataLoader. You'll find helpful functions in the data module of every application to directly create this DataBunch for you.

FastAI.DataBunch — Type

DataBunch
DataBunch(train::Flux.Data.DataLoader, valid::Flux.Data.DataLoader)

A DataBunch is a bunched train and valid (validation) dataloader.