Basic Training Functionality

Basic training wraps together the data (in a DataBunch object) with a Flux model to define a Learner object. The Learner object is the entry point of most of the Callback objects that will customize this training loop in different ways.

Learner

A Learner type wraps a DataBunch with a model and optimizer. The model can then be fit using the optimizer by training it for several epochs. The training process is customized by adding Callbacks to the Learner.

FastAI.Learner — Type

mutable struct Learner <: AbstractLearner

Learner(data_bunch, model; opt = Flux.ADAM(), loss = Flux.mse)

A Learner is the standard grouping of a data bunch, model, optimizer, and loss.

FastAI.add_cb! — Function

add_cb!(learner::Learner, cb::AbstractCallback) -> Array{AbstractCallback,1}

Add cb to the list of callbacks for learner.

FastAI.cbs — Function

cbs(learner::Learner) -> Array{AbstractCallback,N} where N

Get the list of callbacks for learner.

FastAI.data_bunch — Function

data_bunch(l::Learner) -> DataBunch

Get the data bunch for l.

FastAI.data_bunch! — Function

data_bunch!(l::Learner, data_bunch::Any) -> Any

Set the data bunch for l to data_bunch.

FastAI.model — Function

model(l::Learner) -> Any

Get the model for l.

FastAI.model! — Function

model!(l::Learner, model::Any) -> Any

Set the model for l to model.

FastAI.loss — Function

loss(l::Learner) -> Any

Get the loss for l.

FastAI.loss! — Function

loss!(l::Learner, loss::Any) -> Any

Set the loss for l to loss.

FastAI.opt — Function

opt(l::Learner) -> Any

Get the optimizer for l.

FastAI.opt! — Function

opt!(l::Learner, opt::Any) -> Any

Set the optimizer for l to opt.

AbstractLearner Interface

If you want to create a custom learner, you can implement the AbstractLearner interface.

FastAI.AbstractLearner — Type

abstract type AbstractLearner

An AbstractLearner groups together a model, train and validate data, optimizer, loss function and callbacks.

Callbacks are used for every tweak of the training loop. Callbacks receive epoch, batch and loss information which they may pass on to Metrics.

Require interface

data_bunch(l::AbstractLearner)

data_bunch!(l::AbstractLearner, data_bunch)

model(l::AbstractLearner)

model!(l::AbstractLearner, model)

loss(l::AbstractLearner)

loss!(l::AbstractLearner, loss)

opt(l::AbstractLearner)

opt!(l::AbstractLearner, opt)

add_cb!(learner::AbstractLearner, cb::AbstractCallback)

cbs(learner::AbstractLearner)

fit!(learner::AbstractLearner, epoch_count)

Recorder

A Recorder logs the values of metrics throughout the training history.

Missing docstring.

Missing docstring for Recorder. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FastAI.add!(::Recorder, ::Any, ::Any, ::Any). Check Documenter's build log for details.

Missing docstring.

Missing docstring for FastAI.log!. Check Documenter's build log for details.