PyTorch Core Modules

The following modules are PyTorch constructs that fastai builds on, but they don't exist in Julia and Flux.

Dataset

A dataset can either be an IterableDataset or a MapDataset. If you want to create a custom dataset, then you should implement one of these two interfaces.

FastAI.IterableDataset — Type

abstract type IterableDataset

Every dataset is an IterableDataset representing an iterable of data samples. IterableDataset is particularly useful when data come from a stream.

Required interface

iterate(it<:IterableDataset): Returns either a tuple of the first item and initial state or nothing if empty

iterate(it<:IterableDataset, state): Returns either a tuple of the next item and next state or nothing if no items remain

See Julia iteration documentation for more information (https://docs.julialang.org/en/v1/manual/interfaces/)

FastAI.MapDataset — Type

abstract type MapDataset <: IterableDataset

Some datasets are also MapDatasets that map integer ids to data samples. All MapDatasets are also IterableDatasets.

Required interface

Base.getindex(md<:MapDataset, idx::Int): a MapDataset is a indexable type that maps an integer id to a sample. Legal IDs are between 1 and the length of this dataset.

Base.getindex(md<:Dataset, rng::UnitRange): returns a contiguous subset of the items in this dataset

Base.length(md<:MapDataset): returns the number of samples in this MapDataset. Used by many Sampler implementations and the default options of DataLoader.

There are also several dataset types that represent combinations of datasets. If ds1, ds2, ..., dsn are the same kind of dataset, then ds1 ++ ds2 ++ ... ++ dsn will be a combined dataset of that sort. If IterableDatasets and MapDatasets are combined, the result will be an IterableDataset. This is useful to assemble different existing dataset streams. The chainning operation is done on-the-fly, so concatenating large-scale IterableDatasets with this type will be efficient.

FastAI.ConcatDataset — Type

struct ConcatDataset <: MapDataset

ConcatDataset(ds1::MapDataset, ds2::MapDataset)

Two MapDatasets concatenated together.

FastAI.ChainDataset — Type

struct ChainDataset <: IterableDataset

ChainDataset(ds1::IterableDataset, ds2::IterableDataset)

A sequence of IterableDatasets.

FastAI.:++ — Function

++(ds1::MapDataset, ds2::MapDataset) -> FastAI.ConcatDataset

Concatenate two MapDatasets into a ConcatDataset.

++(ds1::IterableDataset, ds2::IterableDataset) -> FastAI.ConcatDataset

Combine two or more IterableDatasets into ChainDataset. Note: if all the datasets are MapDatasets, then the result is a ConcatDataset, but any other combination will result in a ChainDataset.

Lastly, we can also take random and fixed subsets of a MapDataset.

FastAI.SubsetDataset — Type

struct SubsetDataset <: MapDataset

A subset of a MapDataset.

FastAI.subset — Function

subset(dataset::MapDataset, indices::Array{Int64,N} where N) -> FastAI.SubsetDataset

A SubsetDataset of dataset at indices.

FastAI.random_split — Function

random_split(dataset::MapDataset, lengths::Array{Int64,N} where N) -> Array{Any,1}

Randomly split dataset into non-overlapping new SubsetDatasets of given lengths.