Basic Training Functionality

Basic training wraps together the data (in a DataBunch object) with a Flux model to define a Learner object. The Learner object is the entry point of most of the Callback objects that will customize this training loop in different ways.

<!– from fastai docs

Some of the most commonly used customizations are available through the train module, notably:

Learner.lrfind will launch an LR range test that will help you select a good learning rate. Learner.fitonecycle will launch a training using the 1cycle policy to help you train your model faster. Learner.tofp16 will convert your model to half precision and help you launch a training in mixed precision. –>

Learner

A Learner type wraps a DataBunch with a model and optimizer. The model can then be fit using the optimizer by training it for several epochs. The training process is customized by adding Callbacks to the Learner.

FastAI.Learner — Type

Learner <: AbstractLearner
Learner(data_bunch, model; opt = Flux.ADAM(), loss = Flux.mse)

A Learner is the standard grouping of a data bunch, model, optimizer, and loss.

FastAI.add_cb! — Function

add_cb!(learner::Learner, cb::AbstractCallback)

Add cb to the list of callbacks for learner.

Missing docstring.

Missing docstring for cbs. Check Documenter's build log for details.

FastAI.data_bunch — Function

data_bunch(l::Learner)

Get the data bunch for l.

Missing docstring.

Missing docstring for data_bunch!. Check Documenter's build log for details.

FastAI.model — Function

model(l::Learner)

Get the model for l.

Missing docstring.

Missing docstring for model!. Check Documenter's build log for details.

FastAI.loss — Function

loss(l::Learner)

Get the loss for l.

FastAI.loss! — Function

loss!(l::Learner, loss)

Set the loss for l to loss.

FastAI.opt — Function

opt(l::Learner)

Get the optimizer for l.

FastAI.opt! — Function

opt!(l::Learner, opt)

Set the optimizer for l to opt.

AbstractLearner Interface

If you want to create a custom learner, you can implement the AbstractLearner interface.

FastAI.AbstractLearner — Type

AbstractLearner

An AbstractLearner groups together a model, train and validate data, optimizer, loss function and callbacks.

Callbacks are used for every tweak of the training loop. Callbacks receive epoch, batch and loss information which they may pass on to Metrics.

Require interface

data_bunch(l::AbstractLearner)

data_bunch!(l::AbstractLearner, data_bunch)

model(l::AbstractLearner)

model!(l::AbstractLearner, model)

loss(l::AbstractLearner)

loss!(l::AbstractLearner, loss)

opt(l::AbstractLearner)

opt!(l::AbstractLearner, opt)

add_cb!(learner::AbstractLearner, cb::AbstractCallback)

cbs(learner::AbstractLearner)

fit!(learner::AbstractLearner, epoch_count)

Recorder

A Recorder logs the values of metrics throughout the training history.

FastAI.Recorder — Type

Recorder
Recorder(learn::Learner; train_loss = true, train_smooth_loss = true,
                         validate_loss = true, validate_smooth_loss = true)

Container for Learner statistics (e.g. lr, loss and metrics) during training. Statistics are indexed by name, epoch and batch. For example to get the smoothed training loss for epoch 2, batch 3, we would call

recorder["TrainSmoothLoss", 2, 3]

To get the entire history of smooth training loss, one would call

recorder["TrainSmoothLoss", :, :]

Missing docstring.

Missing docstring for add!(::Recorder, ::Any, ::Any, ::Any). Check Documenter's build log for details.

Missing docstring.

Missing docstring for log!. Check Documenter's build log for details.